[#52] Put login routes into their own file in the routes directory
